requirements.txt
python-frontmatter
PyYAML
anthropic
scikit-learn
ollama
